Split up monolithic SConstruct into many small parts.
  * New Module system.
  * Many modules to deal with common libraries.
  * With unit tests for the logic.
