+++ src/beast/python/README.md
@@ -1,0 +1,8 @@
+Unit Tests
+==========
+
+To run beast's Python unit tests, execute:
+
+     python -m unittest discover
+
+from this directory.
